requirements.txt
torchvision==0.5.0
pretrainedmodels==0.7.4
opencv_python==4.2.0.32
easydict==1.9
scipy==1.2.1
pandas==0.24.2
tqdm==4.32.2
numpy==1.16.2
matplotlib==3.0.2
torch==1.4.0
asposestorage==1.0.2
Pillow==7.2.0
scikit_learn==0.23.2
tensorboardX==2.1
torchsummary==1.5.1
transforms==0.1
